+++ apps/web/src/components/Common/Providers/index.tsx
@@ -2,7 +2,6 @@ import { getLivepeerClient, videoPlayerTheme } from '@lenstube/browser'
 import {
   IS_MAINNET,
   LENSTUBE_APP_NAME,
-  LENSTUBE_WEBSITE_URL,
   POLYGON_RPC_URL,
   WC_PROJECT_ID
 } from '@lenstube/constants'
@@ -12,21 +11,16 @@ import { loadLocale } from '@lib/i18n'
 import { i18n } from '@lingui/core'
 import { I18nProvider } from '@lingui/react'
 import { LivepeerConfig } from '@livepeer/react'
-import {
-  connectorsForWallets,
-  darkTheme,
-  lightTheme,
-  RainbowKitProvider
-} from '@rainbow-me/rainbowkit'
-import type { ThemeOptions } from '@rainbow-me/rainbowkit/dist/themes/baseTheme'
+import { connectorsForWallets } from '@rainbow-me/rainbowkit'
 import {
   coinbaseWallet,
   injectedWallet,
   ledgerWallet,
+  metaMaskWallet,
   rainbowWallet,
   walletConnectWallet
 } from '@rainbow-me/rainbowkit/wallets'
-import { ThemeProvider, useTheme } from 'next-themes'
+import { ThemeProvider } from 'next-themes'
 import type { ReactNode } from 'react'
 import React, { useEffect } from 'react'
 import { configureChains, createConfig, WagmiConfig } from 'wagmi'
@@ -34,6 +28,7 @@ import { polygon, polygonMumbai } from 'wagmi/chains'
 import { jsonRpcProvider } from 'wagmi/providers/jsonRpc'
 
 import ErrorBoundary from '../ErrorBoundary'
+import RainbowKit from './RainbowKit'
 
 // TEMP: Duplicate to fix signTypedData_v4 issue. Remove once fixed on WC+WAGMI end.
 const preferredChains = [
@@ -51,9 +46,9 @@ const { chains, publicClient } = configureChains(preferredChains, [
 
 const connectors = connectorsForWallets([
   {
-    groupName: 'Recommended',
+    groupName: 'Popular',
     wallets: [
-      injectedWallet({ chains, shimDisconnect: true }),
+      metaMaskWallet({ chains, projectId: WC_PROJECT_ID }),
       rainbowWallet({ chains, projectId: WC_PROJECT_ID }),
       ledgerWallet({ chains, projectId: WC_PROJECT_ID }),
       coinbaseWallet({ appName: LENSTUBE_APP_NAME, chains }),
@@ -66,7 +61,8 @@ const connectors = connectorsForWallets([
           },
           projectId: WC_PROJECT_ID
         }
-      })
+      }),
+      injectedWallet({ chains, shimDisconnect: true })
     ]
   }
 ])
@@ -76,46 +72,6 @@ const wagmiConfig = createConfig({
   connectors,
   publicClient
 })
-
-const Disclaimer = () => (
-  <div className="prose-sm prose-a:text-indigo-500 text-[10px]">
-    <span>By continuing, you agree to Lenstube's</span>{' '}
-    <a href="/terms" target="_blank">
-      terms
-    </a>{' '}
-    <span>and</span>{' '}
-    <a href="/privacy" target="_blank">
-      privacy policy
-    </a>
-    .
-  </div>
-)
-
-// Enables usage of theme in RainbowKitProvider
-const RainbowKitProviderWrapper = ({ children }: { children: ReactNode }) => {
-  const { theme } = useTheme()
-  const themeOptions: ThemeOptions = {
-    fontStack: 'system',
-    overlayBlur: 'small',
-    accentColor: '#6366f1'
-  }
-  return (
-    <RainbowKitProvider
-      appInfo={{
-        appName: LENSTUBE_APP_NAME,
-        learnMoreUrl: LENSTUBE_WEBSITE_URL,
-        disclaimer: () => <Disclaimer />
-      }}
-      modalSize="compact"
-      chains={chains}
-      theme={
-        theme === 'dark' ? darkTheme(themeOptions) : lightTheme(themeOptions)
-      }
-    >
-      {children}
-    </RainbowKitProvider>
-  )
-}
 
 const Providers = ({ children }: { children: ReactNode }) => {
   useEffect(() => {
@@ -128,11 +84,11 @@ const Providers = ({ children }: { children: ReactNode }) => {
         <LivepeerConfig client={getLivepeerClient()} theme={videoPlayerTheme}>
           <WagmiConfig config={wagmiConfig}>
             <ThemeProvider defaultTheme="dark" attribute="class">
-              <RainbowKitProviderWrapper>
+              <RainbowKit chains={chains}>
                 <ApolloProvider client={apolloClient(authLink)}>
                   {children}
                 </ApolloProvider>
-              </RainbowKitProviderWrapper>
+              </RainbowKit>
             </ThemeProvider>
           </WagmiConfig>
         </LivepeerConfig>
